+++ examples/custom_ui/customui/Button.py
@@ -1,5 +1,5 @@
-from examples.counter.custom_ui.customui.Label import Label
-from examples.counter.custom_ui.customui.base.View import View
+from examples.custom_ui.customui.Label import Label
+from examples.custom_ui.customui.base.View import View
 
 
 class Button(View):
