+++ examples/custom_ui/customui/Label.py
@@ -1,4 +1,4 @@
-from examples.counter.custom_ui.customui.base.View import View
+from examples.custom_ui.customui.base.View import View
 
 
 class Label(View):
